Refactor: VoteRecord의 dto 의존성 제거
- VoteRecord의 toEntity 메소드를 VoteRecordAddRequestDto로 변경하여 의존성 제거
- VoteRecordService의 toEntity 호출 클래스 변경
